Add iterable check and enhance response rendering
The update ensures that the `ResourceObject` body is iterable before processing. It also tweaks the `StreamResponder` to avoid setting HTTP response codes when running under command-line interface (CLI) and only attempts to render `ResourceObject` if its `view` property is `null`. Unnecessary annotation in `IntegrateTest` was also removed and `StreamTransferInject` was updated with necessary dependencies.